Added new controllers, mostly modified model, helper and repository files

+++ Event Management/Repositories/CommentRepositoryFolder/CommentRepository.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Event_Management.Data;
+using Event_Management.Exceptions;
 using Event_Management.Models;
 using Event_Management.Models.Dtos.CommentDtos;
 using Microsoft.EntityFrameworkCore;
@@ -30,7 +31,26 @@ namespace Event_Management.Repositories.CommentRepositoryFolder
             return await _context.Comments
                 .Include(x => x.Event)
                 .Include(x => x.User)
-                .FirstOrDefaultAsync(x => x.Id == id);
+                .FirstOrDefaultAsync(x => x.Id == id)
+                ?? throw new NotFoundException("Comment not found!");
+        }
+
+        public async Task<IEnumerable<Comment>> GetCommentsByEventIdAsync(int eventId)
+        {
+            return await _context.Comments
+                .Where(c => c.EventId == eventId)
+                .Include(x => x.Event)
+                .Include(x => x.User)
+                .ToListAsync();
+        }
+
+        public async Task<IEnumerable<Comment>> GetCommentsByUserIdAsync(int userId)
+        {
+            return await _context.Comments
+                .Where(c => c.UserId == userId)
+                .Include(x => x.Event)
+                .Include(x => x.User)
+                .ToListAsync();
         }
 
         public async Task<Comment> AddCommentAsync(CommentCreateDto commentCreateDto)
@@ -45,10 +65,13 @@ namespace Event_Management.Repositories.CommentRepositoryFolder
             return comment;
         }
 
-        public async Task<bool> UpdateCommentAsync(int id, CommentUpdateDto commentUpdateDto)
+        public async Task<bool> UpdateCommentAsync(int commentId, int userId, CommentUpdateDto commentUpdateDto)
         {
-            var existingComment = await _context.Comments.FindAsync(id);
+            var existingComment = await _context.Comments.FindAsync(commentId);
             if (existingComment == null) return false;
+
+            if (existingComment.UserId == userId)
+                throw new BadRequestException("You don't have permission to update the comment");
 
             _mapper.Map(commentUpdateDto, existingComment);
 
@@ -56,10 +79,13 @@ namespace Event_Management.Repositories.CommentRepositoryFolder
             return true;
         }
 
-        public async Task<bool> DeleteCommentAsync(int id)
+        public async Task<bool> DeleteCommentAsync(int id, int userId)
         {
             var comment = await _context.Comments.FindAsync(id);
             if (comment == null) return false;
+
+            if (comment.UserId == userId)
+                throw new BadRequestException("You don't have permission to delete the comment");
 
             _context.Comments.Remove(comment);
             await _context.SaveChangesAsync();
