+++ Event Management/Repositories/ReviewRepositoryFolder/ReviewRepository.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Event_Management.Data;
+using Event_Management.Exceptions;
 using Event_Management.Models;
 using Event_Management.Models.Dtos.ReviewDtos;
 using Microsoft.EntityFrameworkCore;
@@ -30,7 +31,26 @@ namespace Event_Management.Repositories.ReviewRepositoryFolder
             return await _context.Reviews
                 .Include(x => x.Event)
                 .Include(x => x.User)
-                .FirstOrDefaultAsync(x => x.Id == id);
+                .FirstOrDefaultAsync(x => x.Id == id)
+                ?? throw new NotFoundException("Review not found!");
+        }
+
+        public async Task<IEnumerable<Review>> GetReviewsByEventIdAsync(int eventId)
+        {
+            return await _context.Reviews
+                .Where(r => r.EventId == eventId)
+                .Include(x => x.Event)
+                .Include(x => x.User)
+                .ToListAsync();
+        }
+
+        public async Task<IEnumerable<Review>> GetReviewsByUserIdAsync(int userId)
+        {
+            return await _context.Reviews
+                .Where(r => r.UserId == userId)
+                .Include(x => x.Event)
+                .Include(x => x.User)
+                .ToListAsync();
         }
 
         public async Task<Review> AddReviewAsync(ReviewCreateDto reviewCreateDto)
@@ -45,10 +65,13 @@ namespace Event_Management.Repositories.ReviewRepositoryFolder
             return review;
         }
 
-        public async Task<bool> UpdateReviewAsync(int id, ReviewUpdateDto reviewUpdateDto)
+        public async Task<bool> UpdateReviewAsync(int reviewId, int userId, ReviewUpdateDto reviewUpdateDto)
         {
-            var existingReview = await _context.Reviews.FindAsync(id);
+            var existingReview = await _context.Reviews.FindAsync(reviewId);
             if (existingReview == null) return false;
+
+            if(existingReview.UserId == userId) 
+                throw new BadRequestException("You don't have permission to update the review!");
 
             _mapper.Map(reviewUpdateDto, existingReview);
 
@@ -56,10 +79,13 @@ namespace Event_Management.Repositories.ReviewRepositoryFolder
             return true;
         }
 
-        public async Task<bool> DeleteReviewAsync(int id)
+        public async Task<bool> DeleteReviewAsync(int reviewId, int userId)
         {
-            var review = await _context.Reviews.FindAsync(id);
+            var review = await _context.Reviews.FindAsync(reviewId);
             if (review == null) return false;
+
+            if (review.UserId == userId)
+                throw new BadRequestException("You don't have permission to delete the review!");
 
             _context.Reviews.Remove(review);
             await _context.SaveChangesAsync();
